kinda draws letters? badly

+++ PhobGCC/common/phobGCC.h
@@ -1999,7 +1999,7 @@ void readSticks(int readA, int readC, Buttons &btn, Pins &pin, const Buttons &ha
 	//Read the sticks repeatedly until it's been 1 millisecond since the last iteration
 	//This is for denoising and making sure the loop runs at 1000 Hz
 	static uint64_t lastMicros = micros();
-	static int adcDelta = 0;
+	static uint64_t adcDelta = 0;
 	uint64_t beforeMicros = micros();
 	uint64_t afterMicros;
 
@@ -2014,7 +2014,7 @@ void readSticks(int readA, int readC, Buttons &btn, Pins &pin, const Buttons &ha
 		adcDelta = afterMicros-beforeMicros;
 		beforeMicros = afterMicros;
 	}
-	while(int64_t(afterMicros-lastMicros) < (1000 - adcDelta));
+	while(uint64_t(afterMicros-lastMicros) < (1000 - adcDelta));
 
 	//Then we spinlock to get the 1 kHz more exactly.
 	while((afterMicros-lastMicros) < 1000) {
